Windows build?

+++ src/util/handle.h
@@ -145,7 +145,7 @@ private:
 
         assert(oldSize < Handle::MAX_INDICES && "Max table capacity reached!");
 
-        std::size_t newSize = std::min(oldSize + grow, Handle::MAX_INDICES);
+        std::size_t newSize = std::min(oldSize + grow, (size_t)Handle::MAX_INDICES);
 
         this->m_Table.resize(newSize);
